prepare/reorg code for switching to use install user creds when doing refreshes
